requirements.txt
flask
gTTS
pyttsx3
google-cloud-texttospeech
pydub
